Use right variable containing URI object when opening website. Also moved functionality to NammuController so it's accessible from both Toolbar and Menu options.
